requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pip==22.0.4
python-dotenv==0.19.2
requests==2.27.1
setuptools==58.5.3
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.37.0
